Create sethares-dissonance.py
Compute a Plomp-Levelt dissonance curve for a sound that is specified in terms of frequencies and amplitudes of partials. The algorithm and parameters are due to Will Sethares. This is based on C code by Morgan Thunder. User endolith has already made a python version, but I wanted one without a numpy dependency.